refector: separate TripDatePicker from App
