+++ app/src/TripDatePicker.css
@@ -1,0 +1,14 @@
+.date-picker-content {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  z-index: 100;
+  background-color: var(--color-tropical-coconut-white-sand); 
+  border-radius: 5px;
+  box-shadow:  0 4px 8px rgba(0, 0, 0, 0.1);
+  background-color: #FFFFFF;
+  fill: white;
+}
